data: add packets to queue even if init fails
Use a work item to go back and add a keypair/nonces when receiving a
handshake response. This removes the need (on the parallel path) to
place packets back onto tx_packet_queue.

We must be careful to handle all of the cases where a new keypair is
used, by running our work function then.

Signed-off-by: Samuel Holland <samuel@sholland.org>